doc: add minimal Checkstyle configuration example
