+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/missingjavadocmethod/InputMissingJavadocMethodNoJavadoc2B.java
@@ -16,56 +16,7 @@ package com.puppycrawl.tools.checkstyle.checks.javadoc.missingjavadocmethod;
 /* Config:
  * scope = "protected"
  */
-public class InputMissingJavadocMethodNoJavadoc2 //comment test
-{
-    public int i1;
-    protected int i2;
-    int i3;
-    private int i4;
-
-    public void foo1() {} // violation
-    protected void foo2() {} // violation
-    void foo3() {}
-    private void foo4() {}
-
-    protected class ProtectedInner {
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {} // violation
-        protected void foo2() {} // violation
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    class PackageInner {
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    private class PrivateInner {
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-}
-
-class PackageClass2 {
+class InputMissingJavadocMethodNoJavadoc2B {
     public int i1;
     protected int i2;
     int i3;
