+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/missingjavadocmethod/InputMissingJavadocMethodNoJavadocB.java
@@ -16,56 +16,7 @@ package com.puppycrawl.tools.checkstyle.checks.javadoc.missingjavadocmethod;
 /* Config:
  * scope = "private"
  */
-public class InputMissingJavadocMethodNoJavadoc //comment test
-{
-    public int i1;
-    protected int i2;
-    int i3;
-    private int i4;
-
-    public void foo1() {} // violation
-    protected void foo2() {} // violation
-    void foo3() {} // violation
-    private void foo4() {} // violation
-
-    protected class ProtectedInner {
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {} // violation
-        protected void foo2() {} // violation
-        void foo3() {} // violation
-        private void foo4() {} // violation
-    }
-
-    class PackageInner {
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {} // violation
-        protected void foo2() {} // violation
-        void foo3() {} // violation
-        private void foo4() {} // violation
-    }
-
-    private class PrivateInner {
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {} // violation
-        protected void foo2() {} // violation
-        void foo3() {} // violation
-        private void foo4() {} // violation
-    }
-}
-
-class PackageClass {
+class InputMissingJavadocMethodNoJavadocB {
     public int i1;
     protected int i2;
     int i3;
